+++ src/vox.c
@@ -24,7 +24,7 @@
 static const uint32_t VOX_DEFAULT_PALETTE[256];
 
 #define READ(type, file) \
-    ({ type v; int r = fread(&v, sizeof(v), 1, file); (void)r; v;})
+    ({ type v; size_t r = fread(&v, sizeof(v), 1, file); (void)r; v;})
 #define WRITE(type, v, file) \
     ({ type v_ = v; fwrite(&v_, sizeof(v_), 1, file);})
